refactor(specialists): adjust nomenclature to plural in files and classes
