+++ src/app/http/specialists/specialists.module.ts
@@ -1,14 +1,14 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { Specialist } from '@/domain/entities/specialists';
+import { Specialist } from '@/domain/entities/specialist';
 
-import { SpecialistController } from './specialists.controller';
+import { SpecialistsController } from './specialists.controller';
 import { SpecialistsRepository } from './specialists.repository';
 import { SpecialistsService } from './specialists.service';
 @Module({
   imports: [TypeOrmModule.forFeature([Specialist])],
-  controllers: [SpecialistController],
+  controllers: [SpecialistsController],
   providers: [SpecialistsService, SpecialistsRepository],
   exports: [SpecialistsRepository],
 })
