+++ src/app/http/specialists/specialists.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
-import { Specialist } from '@/domain/entities/specialists';
+import { Specialist } from '@/domain/entities/specialist';
 
 @Injectable()
 export class SpecialistsRepository {
